+++ app/src/main/java/com/homework/uilistshomework/ThirdViewHolder.kt
@@ -4,17 +4,28 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.homework.uilistshomework.databinding.ItemTabFirstBinding
 
-class ThirdViewHolder(private val binding: ItemTabFirstBinding) : BaseViewHolder(binding.root) {
+class ThirdViewHolder(private val binding: ItemTabFirstBinding,
+                      private val clickRemoveItem : IchangeList) : BaseViewHolder(binding.root) {
 
-    constructor(parent: ViewGroup) : this(
-            ItemTabFirstBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+    constructor(parent: ViewGroup, clickRemoveItem: IchangeList) : this(
+        ItemTabFirstBinding.inflate(LayoutInflater.from(parent.context), parent, false),
+        clickRemoveItem
     )
 
-    fun bind(item: Item, click: Adapter.Idelete, position: Int) {
+    fun bind(item: Item) {
         binding.itemText.setBackgroundResource(item.shape)
         binding.itemText.text = item.name
-        binding.deleteItem.setOnClickListener {
-            click.removeListItem(position)
+        addListener(item)
+    }
+
+    private var item: Item? = null
+
+    override fun addListener(item: Item) {
+        if (this.item == null) {
+            binding.deleteItem.setOnClickListener {
+                clickRemoveItem.removeItemUpdate(item)
+            }
+            this.item = item
         }
     }
 }
